requirements.txt
numpy==1.11.2
pika==0.10.0
redis==2.10.5
Pillow==3.4.2
